+++ pico-cnn/layers/convolution.c
@@ -1,7 +1,7 @@
 #include "convolution.h"
 
-void convolution1d_naive(const fp_t* input_channel, const int input_size, fp_t* output_channel, const fp_t* kernel,
-                         const int kernel_size, const int stride, const int padding, const fp_t bias) {
+void convolution1d_naive(const fp_t* input_channel, const uint16_t input_size, fp_t* output_channel, const fp_t* kernel,
+                         const uint16_t kernel_size, const uint16_t stride, const uint16_t padding, const fp_t bias) {
 
     int32_t input_channel_idx;
     int32_t kernel_idx;
@@ -52,7 +52,7 @@ void convolution1d_naive(const fp_t* input_channel, const int input_size, fp_t* 
 void convolution2d_padding_naive(const fp_t* input_channel, const uint16_t height, const uint16_t width, fp_t* output_channel,
                                  const fp_t* kernel, const uint16_t kernel_height, const uint16_t kernel_width,
                                  const uint16_t stride_height, const uint16_t stride_width,
-                                 const int* padding, const fp_t bias) {
+                                 const uint16_t* padding, const fp_t bias) {
 
    fp_t* new_input_channel;
