+++ pico-cnn/layers/fully_connected.c
@@ -2,7 +2,7 @@
 
 void fully_connected_naive(const fp_t* input_channel, const uint16_t input_width, fp_t* output_channel, const uint16_t output_width, const fp_t* kernel, const fp_t* bias) {
 
-    int i, j;
+    uint16_t i, j;
     for(i = 0; i < output_width; i++) {
 
         fp_t pixel = 0.0;
