+++ test/layers/test_pooling.c
@@ -16,9 +16,9 @@ void pooling_output_channel_dimensions(uint16_t height, uint16_t width, const ui
     (*computed_dimensions)[1] = output_channel_width;
 }
 
-int test_max_pooling1d() {
+int32_t test_max_pooling1d() {
     printf("test_max_pooling1d()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     uint16_t input_height = 1;
     uint16_t input_width = 16;
@@ -42,7 +42,7 @@ int test_max_pooling1d() {
 
     max_pooling1d_naive(input, input_width, output, kernel_width, stride);
 
-    for(int i = 0; i < expected_output_width; i++) {
+    for(int32_t i = 0; i < expected_output_width; i++) {
         if(output[i] != expected_output[i]){
             printf("Expected: %f, Output: %f\n", expected_output[i], output[i]);
             return_value = 1;
@@ -52,9 +52,9 @@ int test_max_pooling1d() {
     return return_value;
 }
 
-int test_max_pooling1d_padding() {
+int32_t test_max_pooling1d_padding() {
     printf("test_max_pooling1d_padding()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     uint16_t input_height = 1;
     uint16_t input_width = 16;
@@ -63,7 +63,7 @@ int test_max_pooling1d_padding() {
     uint16_t kernel_height = 1;
     uint16_t kernel_width = 3;
     uint16_t stride = 2;
-    const int padding[2] = {1, 1};
+    const uint16_t padding[2] = {1, 1};
     uint16_t kernel_dim[2] = {kernel_height, kernel_width};
 
     fp_t input[16] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16};
@@ -80,7 +80,7 @@ int test_max_pooling1d_padding() {
 
     max_pooling1d_naive_padded(input, input_width, output, kernel_width, stride, padding);
 
-    for(int i = 0; i < expected_output_width; i++) {
+    for(int32_t i = 0; i < expected_output_width; i++) {
         if(expected_output[i] != output[i]){
             printf("Expected: %f, Output; %f\n", expected_output[i], output[i]);
             return_value = 1;
@@ -90,9 +90,9 @@ int test_max_pooling1d_padding() {
     return return_value;
 }
 
-int test_max_pooling2d() {
+int32_t test_max_pooling2d() {
     printf("test_max_pooling2d()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     fp_t input[16] = {1, 2, 3, 4,
                       5, 6, 7, 8,
@@ -106,7 +106,7 @@ int test_max_pooling2d() {
 
     max_pooling2d_naive(input, 4, 4, output, 3, 1);
 
-    for(int i = 0; i < 4; i++) {
+    for(int32_t i = 0; i < 4; i++) {
         if(output[i] != expected_output[i]) {
             printf("Expected: %f, Output; %f\n", expected_output[i], output[i]);
             return_value = 1;
@@ -116,9 +116,9 @@ int test_max_pooling2d() {
     return return_value;
 }
 
-int test_max_pooling2d_padding() {
+int32_t test_max_pooling2d_padding() {
     printf("test_max_pooling2d_padding()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     fp_t input[25] = {1, 2, 3, 4, 5,
                       6, 7, 8, 9, 10,
@@ -134,10 +134,10 @@ int test_max_pooling2d_padding() {
 
     fp_t* output = (fp_t*) malloc(25* sizeof(float));
 
-    const int padding[4] = {2, 2, 2, 2};
+    const uint16_t padding[4] = {2, 2, 2, 2};
     max_pooling2d_naive_padded(input, 5, 5, output, 5, 1, padding);
 
-    for(int i = 0; i < 25; i++) {
+    for(int32_t i = 0; i < 25; i++) {
         if(output[i] != expected_output[i]) {
             printf("Expected: %f, Output; %f\n", expected_output[i], output[i]);
             return_value = 1;
@@ -147,9 +147,9 @@ int test_max_pooling2d_padding() {
     return return_value;
 }
 
-int test_avg_pooling1d() {
+int32_t test_avg_pooling1d() {
     printf("test_avg_pooling1d()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     uint16_t input_width = 10;
     uint16_t expected_output_width = 8;
@@ -173,7 +173,7 @@ int test_avg_pooling1d() {
     average_pooling1d_naive(input, input_width, output, kernel_size, stride,
                             0.0, count_include_pad);
 
-    for(int i = 0; i < 1*expected_output_width; i++){
+    for(int32_t i = 0; i < 1*expected_output_width; i++){
         if(expected_output[i] != output[i]) {
             printf("Expected: %f, Output; %f\n", expected_output[i], output[i]);
             return_value = 1;
@@ -184,9 +184,9 @@ int test_avg_pooling1d() {
     return return_value;
 
 }
-int test_avg_pooling1d_padding() {
+int32_t test_avg_pooling1d_padding() {
     printf("test_avg_pooling1d_padding()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     uint16_t input_width = 10;
     uint16_t expected_output_width = 10;
@@ -202,14 +202,14 @@ int test_avg_pooling1d_padding() {
     fp_t* output_0 = (fp_t*) malloc(expected_output_width* sizeof(float));
     fp_t* output_1 = (fp_t*) malloc(expected_output_width* sizeof(float));
 
-    const int padding[4] = {2, 2, 2, 2};
+    const uint16_t padding[4] = {2, 2, 2, 2};
 
     average_pooling1d_naive_padded(input, input_width, output_0, kernel_size,
                                    stride, 0.0, padding, 0);
     average_pooling1d_naive_padded(input, input_width, output_1, kernel_size,
                                    stride, 0.0, padding, 1);
 
-    for(int i = 0; i < expected_output_width; i++) {
+    for(int32_t i = 0; i < expected_output_width; i++) {
         if (expected_output_count_include_pad_0[i] != output_0[i]) {
             return_value = 1;
         }
@@ -224,9 +224,9 @@ int test_avg_pooling1d_padding() {
     return return_value;
 }
 
-int test_avg_pooling2d() {
+int32_t test_avg_pooling2d() {
     printf("test_avg_pooling2d()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     uint16_t input_height = 5;
     uint16_t input_width = 5;
@@ -259,7 +259,7 @@ int test_avg_pooling2d() {
     average_pooling2d_naive(input, input_height, input_width, output, kernel_size, stride,
                             0.0, count_include_pad);
 
-    for(int i = 0; i < expected_output_height*expected_output_width; i++){
+    for(int32_t i = 0; i < expected_output_height*expected_output_width; i++){
         if(expected_output[i] != output[i]) {
             printf("Expected: %f, Output; %f\n", expected_output[i], output[i]);
             return_value = 1;
@@ -270,9 +270,9 @@ int test_avg_pooling2d() {
     return return_value;
 }
 
-int test_avg_pooling2d_padding() {
+int32_t test_avg_pooling2d_padding() {
     printf("test_avg_pooling2d_padding()\n");
-    int return_value = 0;
+    int32_t return_value = 0;
 
     uint16_t input_height = 5;
     uint16_t input_width = 5;
@@ -309,18 +309,18 @@ int test_avg_pooling2d_padding() {
     fp_t* output_1 = (fp_t*) malloc(25* sizeof(float));
     fp_t* output_2 = (fp_t*) malloc(25* sizeof(float));
 
-    const int padding[4] = {2, 2, 2, 2};
+    const uint16_t padding[4] = {2, 2, 2, 2};
 
     average_pooling2d_naive_padded(input, input_height, input_width, output_0, kernel_size,
                                    stride, 0.0, padding, 0);
     average_pooling2d_naive_padded(input, input_height, input_width, output_1, kernel_size,
                                    stride, 0.0, padding, 1);
 
-    const int padding2[4] = {1, 1, 1, 1};
+    const uint16_t padding2[4] = {1, 1, 1, 1};
     average_pooling2d_naive_padded(input, input_height, input_width, output_2, 3,
                                    stride, 0.0, padding2, 0);
 
-    for(int i = 0; i < 25; i++){
+    for(int32_t i = 0; i < 25; i++){
         if(expected_output_count_include_pad_0[i] != output_0[i]) {
             return_value = 1;
             printf("Include pad 0: Index: %d: Expected: %f, Output; %f\n", i, expected_output_count_include_pad_0[i], output_0[i]);
@@ -343,9 +343,9 @@ int test_avg_pooling2d_padding() {
     return return_value;
 }
 
-int test_global_average_pooling2d(){
+int32_t test_global_average_pooling2d(){
 
-    int return_value = 0;
+    int32_t return_value = 0;
     printf("test_global_average_pooling2d()\n");
 
     #define input_width 4
@@ -376,9 +376,9 @@ int test_global_average_pooling2d(){
 
 }
 
-int test_global_max_pooling2d(){
+int32_t test_global_max_pooling2d(){
 
-    int return_value = 0;
+    int32_t return_value = 0;
     printf("test_global_max_pooling2d()\n");
 
     #define input_width 5
